src/requirements.txt
Adafruit-Blinka==6.4.1
adafruit-circuitpython-neopixel==6.0.2
adafruit-circuitpython-pypixelbuf==2.2.4
Adafruit-PlatformDetect==3.4.1
Adafruit-PureIO==1.1.8
astral==2.2
board==1.0
pkg-resources==0.0.0
pyftdi==0.52.9
pyserial==3.5
pytz==2021.1
pyusb==1.1.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
